+++ src/main/java/com/example/clinica/model/entities/City.java
@@ -1,19 +1,19 @@
 package com.example.clinica.model.entities;
 
 public class City {
-    private Integer id;
+    private Long id;
     private String name;
 
-    public City(Integer id, String name) {
+    public City(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
